+++ R/02-importData.R
@@ -18,9 +18,11 @@ importDataUI <- function(id, label = "Import Data") {
 #' @param id namespace id
 #' @param title title of data import module
 #' @param defaultSource (character) default source for input "Source", e.g. "ckan", "file", or "url"
-#' @param ckanFileTypes (character) file types allowed for import from Pandora ("ckan")
+#' @param ckanFileTypes (character) file types allowed for import from Pandora ("ckan"). E.g. for
+#' `importType = "data"`: c("xls", "xlsx", "csv", "odt", "txt"); for `importType = "zip"`: c("zip");
+#'  for `importType = "list"`: c("json")
 #' @param ignoreWarnings TRUE to enable imports in case of warnings
-#' @param importType (character) type of import, either "data" or "model" or "zip".
+#' @param importType (character) type of import, either "data", "model", "zip" or "list".
 #'  ImportType == "model" expects a zip file containing a model. The file will be unzipped,
 #'  the model object extracted, and checked if it is valid for the app.
 #'  ImportType == "zip" enables the optional parameter 'expectedFileInZip'. The file is validated
@@ -60,7 +62,7 @@ importDataServer <- function(id,
                              defaultSource = c("ckan", "file", "url", "remoteModel"),
                              ckanFileTypes = c("xls", "xlsx", "csv", "odt", "txt"),
                              ignoreWarnings = FALSE,
-                             importType = c("data", "model", "zip"),
+                             importType = c("data", "model", "zip", "list"),
                              # parameters for data upload
                              rowNames = reactiveVal(NULL),
                              colNames = reactiveVal(NULL),
@@ -88,6 +90,8 @@ importDataServer <- function(id,
   moduleServer(id,
                function(input, output, session) {
                  ns <- session$ns
+                 logDebug(initServerLogTxt(ns("")))
+
                  mergeList <- reactiveVal(list())
                  customNames <- reactiveValues(
                    withRownames = FALSE,
@@ -492,7 +496,8 @@ importDataDialog <-
       title <- switch(importType,
                       "data" = "Data import",
                       "model" = "Model import",
-                      "zip" = "Zip import")
+                      "zip" = "Zip import",
+                      "list" = "Json import")
     }
 
     modalDialog(