+++ NEWS.md
@@ -1,3 +1,10 @@
+# DataTools 24.07.0
+
+## New Features
+- _Import Data_: 
+  - option to import a list from a json file from url, Pandora or file (#91)
+  - option to import user inputs together with a json file in order to load a previous session (#91)
+
 # DataTools 24.06.0
 
 ## New Features